test(integration): move controller batch
